Added a 25Hz timer tick on FIRQ, using the 88C681
This requires initing the counter/timer in crystal/16 mode, setting the
counter starting value to 4605, since: 25 * (4605*2) * 16 = 3.684Mhz, and
turning on interrupts for the coutner/timer.

Uptime is only counted outside the monitor, since everything but NMI
is ignored in the software interrupt.

Also added a U comand to clear the uptime, which is handy when testing.

Oh: and a R commnd for reading an individual byte (was useful when
getting this working). This will go away when the d command can be used
on arbitary addresses with arbitary counts.