Added "IO" channels, kind of.
All character IO is now done through io* instad of serial* functions.
It is possible to select if serial or v99+keyboard should be used.

Booting now assumes v99+keyboard, but if you hit space on serial port that
is used instead.

Changed the jump table so the old serial* names point at the io* functions,
which means I don't have to copy updated user code to the CF. :)

Also tided up the timer and put it in it's own file.